+++ src/main/java/com/example/catchtable/repository/WaitingRepository.java
@@ -28,11 +28,11 @@ public class WaitingRepository {
 
   private final RowMapper<Waiting> waitingRowMapper = (rs, rowNum) ->
       Waiting.fromEntity(
-          rs.getInt("id"),
+          rs.getLong("id"),
           rs.getTimestamp("created_at"),
-          rs.getInt("customer_id"),
+          rs.getLong("customer_id"),
           rs.getInt("guest_count"),
-          rs.getInt("restaurant_id")
+          rs.getLong("restaurant_id")
       );
 
   public Optional<Waiting> save(Waiting entity) {
@@ -48,13 +48,13 @@ public class WaitingRepository {
     KeyHolder keyHolder = new GeneratedKeyHolder();
     jdbcTemplate.update(con -> {
       PreparedStatement ps = con.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-      ps.setInt(1, entity.getCustomerId());
-      ps.setInt(2, entity.getGuestCount());
-      ps.setInt(3, entity.getRestaurantId());
+      ps.setLong(1, entity.getCustomerId());
+      ps.setLong(2, entity.getGuestCount());
+      ps.setLong(3, entity.getRestaurantId());
       return ps;
     }, keyHolder);
     Number key = keyHolder.getKey();
-    return findById(Objects.requireNonNull(key).intValue());
+    return findById(Objects.requireNonNull(key).longValue());
   }
 
   private Optional<Waiting> update(Waiting entity) {
@@ -68,13 +68,13 @@ public class WaitingRepository {
     return findAll(entities);
   }
 
-  public Optional<Waiting> findById(Integer id) {
+  public Optional<Waiting> findById(Long id) {
     String sql = "SELECT * FROM waiting WHERE id = ?";
     List<Waiting> result = jdbcTemplate.query(sql, waitingRowMapper, id);
     return result.isEmpty() ? Optional.empty() : Optional.of(result.get(0));
   }
 
-  public boolean existsById(Integer id) {
+  public boolean existsById(Long id) {
     String sql = "SELECT count(*) FROM waiting WHERE id = ?";
     var result = jdbcTemplate.queryForObject(sql, Long.class, id);
     return Optional.ofNullable(result).orElse(0L) > 0;
@@ -101,7 +101,7 @@ public class WaitingRepository {
     return Optional.ofNullable(result).orElse(0L);
   }
 
-  public void deleteById(Integer id) {
+  public void deleteById(Long id) {
     String sql = "DELETE FROM waiting WHERE id = ?";
     jdbcTemplate.update(sql, id);
   }
